// src/app/api/models/time-entry.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Project } from '../models/project';
import { User } from '../models/user';
export interface TimeEntry {
  breakMinutes?: number;
  comment?: string | null;
  date?: string;
  id?: string;
  minutesWorked?: number;
  project?: Project;
  projectId?: string;
  user?: User;
  userId?: string;
}